refactor: BankManager 타입 정의, 코드 컨벤션 수정
